# tokpulse-railway-fixed/tokpulse-build/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TokPulse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface OAuthUrlResponse {
  url: string;
}

export type UserProfilePlan = typeof UserProfilePlan[keyof typeof UserProfilePlan];


export const UserProfilePlan = {
  free: 'free',
  premium: 'premium',
  grace: 'grace',
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  niche?: string | null;
  plan: UserProfilePlan;
  /** @nullable */
  premiumExpiresAt?: string | null;
  createdAt: string;
}

export type UserProfileUpdateCurrency = typeof UserProfileUpdateCurrency[keyof typeof UserProfileUpdateCurrency];


export const UserProfileUpdateCurrency = {
  NGN: 'NGN',
  USD: 'USD',
} as const;

export interface UserProfileUpdate {
  displayName?: string;
  country?: string;
  currency?: UserProfileUpdateCurrency;
  niche?: string;
}

export interface TikTokAccount {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  followerCount: number;
  followingCount: number;
  videoCount: number;
  /** @nullable */
  lastSyncedAt?: string | null;
}

export interface TikTokConnectInput {
  code: string;
  redirectUri?: string;
}

export interface AnalyticsOverview {
  totalLikes: number;
  totalComments: number;
  totalShares: number;
  totalViews: number;
  totalVideos: number;
  avgEngagementRate: number;
  followerCount?: number;
  /** @nullable */
  lastUpdatedAt: string | null;
}

export interface VideoAnalytics {
  id: number;
  videoId: string;
  title: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  likes: number;
  comments: number;
  shares: number;
  views: number;
  playCount: number;
  engagementRate?: number;
  createdAt?: string;
}

export interface GrowthDataPoint {
  date: string;
  followers: number;
  likes: number;
  comments: number;
  shares: number;
}

export type PaymentInitInputCurrency = typeof PaymentInitInputCurrency[keyof typeof PaymentInitInputCurrency];


export const PaymentInitInputCurrency = {
  NGN: 'NGN',
  USD: 'USD',
} as const;

export interface PaymentInitInput {
  currency: PaymentInitInputCurrency;
}

export type PaymentRequestStatus = typeof PaymentRequestStatus[keyof typeof PaymentRequestStatus];


export const PaymentRequestStatus = {
  pending: 'pending',
  proof_submitted: 'proof_submitted',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface PaymentRequest {
  id: number;
  status: PaymentRequestStatus;
  amount: number;
  currency: string;
  /** @nullable */
  proofImageUrl?: string | null;
  createdAt: string;
}

export interface PaymentProofInput {
  paymentId: number;
  proofImageBase64: string;
  fileName: string;
}

export type SubscriptionPlan = typeof SubscriptionPlan[keyof typeof SubscriptionPlan];


export const SubscriptionPlan = {
  free: 'free',
  premium: 'premium',
  grace: 'grace',
} as const;

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  active: 'active',
  expired: 'expired',
  grace_period: 'grace_period',
  none: 'none',
} as const;

export interface Subscription {
  plan: SubscriptionPlan;
  status: SubscriptionStatus;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  daysRemaining?: number | null;
  latestPayment?: PaymentRequest;
}

export type AiGenerateInputType = typeof AiGenerateInputType[keyof typeof AiGenerateInputType];


export const AiGenerateInputType = {
  caption: 'caption',
  hashtags: 'hashtags',
  viral_ideas: 'viral_ideas',
  content_plan: 'content_plan',
} as const;

export interface AiGenerateInput {
  type: AiGenerateInputType;
  topic?: string;
  videoTitle?: string;
  niche?: string;
}

export interface AiGenerateResponse {
  result: string;
  suggestions?: string[];
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AiChatInput {
  message: string;
  history?: ChatMessage[];
}

export interface AiChatResponse {
  reply: string;
}

export interface PushSubscriptionInput {
  endpoint: string;
  p256dh: string;
  auth: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminInfo {
  email: string;
  name: string;
}

export interface AdminAuthResult {
  token: string;
  admin: AdminInfo;
}

export interface AdminPaymentDetail {
  id: number;
  userId: string;
  userEmail: string;
  /** @nullable */
  userName?: string | null;
  amount: number;
  currency: string;
  status: string;
  /** @nullable */
  proofImageUrl?: string | null;
  createdAt: string;
}

export interface AdminUserDetail {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  plan: string;
  /** @nullable */
  premiumExpiresAt?: string | null;
  tiktokConnected?: boolean;
  createdAt: string;
}

export interface TogglePremiumInput {
  premium: boolean;
  durationDays?: number;
}

export interface AdminStats {
  totalUsers: number;
  premiumUsers: number;
  freeUsers: number;
  pendingPayments: number;
  approvedPayments: number;
  totalRevenue: number;
}

export type GetGrowthChartParams = {
period?: GetGrowthChartPeriod;
};

export type GetGrowthChartPeriod = typeof GetGrowthChartPeriod[keyof typeof GetGrowthChartPeriod];


export const GetGrowthChartPeriod = {
  '7d': '7d',
  '30d': '30d',
  '90d': '90d',
} as const;