# tokpulse-railway-fixed/tokpulse-build/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TokPulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get analytics overview (premium only)
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
        "403":
          description: Premium required
  /analytics/videos:
    get:
      operationId: getVideoAnalytics
      tags: [analytics]
      summary: Get per-video analytics (premium only)
      responses:
        "200":
          description: Video analytics list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoAnalytics"
        "403":
          description: Premium required
  /analytics/growth:
    get:
      operationId: getGrowthChart
      tags: [analytics]
      summary: Get follower/engagement growth chart data (premium only)
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [7d, 30d, 90d]
          required: false
      responses:
        "200":
          description: Growth chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrowthDataPoint"
...
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate a premium subscription payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitInput"
      responses:
        "201":
          description: Payment request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
  /payments/upload-proof:
    post:
      operationId: uploadPaymentProof
      tags: [payments]
      summary: Upload payment proof screenshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "200":
          description: Proof submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResult"
        "401":
          description: Invalid credentials
...
  /admin/payments/{paymentId}/approve:
    post:
      operationId: approvePayment
      tags: [admin]
      summary: Approve a payment and grant premium access
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/payments/{paymentId}/reject:
    post:
      operationId: rejectPayment
      tags: [admin]
      summary: Reject a payment
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/users/{userId}/toggle-premium:
    post:
      operationId: toggleUserPremium
      tags: [admin]
      summary: Manually grant or revoke premium access
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TogglePremiumInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    AnalyticsOverview:
      type: object
      required: [totalLikes, totalComments, totalShares, totalViews, totalVideos, avgEngagementRate, lastUpdatedAt]
      properties:
        totalLikes:
          type: integer
        totalComments:
          type: integer
        totalShares:
          type: integer
        totalViews:
          type: integer
        totalVideos:
          type: integer
        avgEngagementRate:
          type: number
        followerCount:
          type: integer
        lastUpdatedAt:
          type: ["string", "null"]
    VideoAnalytics:
      type: object
      required: [id, videoId, title, likes, comments, shares, views, playCount]
      properties:
        id:
          type: integer
        videoId:
          type: string
        title:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        likes:
          type: integer
        comments:
          type: integer
        shares:
          type: integer
        views:
          type: integer
        playCount:
          type: integer
        engagementRate:
          type: number
        createdAt:
          type: string
    GrowthDataPoint:
      type: object
      required: [date, followers, likes, comments, shares]
      properties:
        date:
          type: string
        followers:
          type: integer
        likes:
          type: integer
        comments:
          type: integer
        shares:
          type: integer
    PaymentInitInput:
      type: object
      required: [currency]
      properties:
        currency:
          type: string
          enum: [NGN, USD]
    PaymentRequest:
      type: object
      required: [id, status, amount, currency, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, proof_submitted, approved, rejected]
        amount:
          type: number
        currency:
          type: string
        proofImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentProofInput:
      type: object
      required: [paymentId, proofImageBase64, fileName]
      properties:
        paymentId:
          type: integer
        proofImageBase64:
          type: string
        fileName:
          type: string
...
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AdminAuthResult:
      type: object
      required: [token, admin]
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminInfo"
    AdminInfo:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
        name:
          type: string
...
    TogglePremiumInput:
      type: object
      required: [premium]
      properties:
        premium:
          type: boolean
        durationDays:
          type: integer
...