# tokpulse-railway-fixed/tokpulse-build/lib/api-zod/src/generated/types/adminAuthResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TokPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminInfo } from './adminInfo';

export interface AdminAuthResult {
  token: string;
  admin: AdminInfo;
}